Add stream en/decoder, rename `-b` flag to `-m`
Flag renamed because `-m` appears to be more standard and is what the
reference implementation uses. The stream coders are inspired by
stdlib's compression API and should make it much easier to work with
streaming data. Some small changes to the buffered en/decoders to make
it work better with the stream wrapper